my_site/requirements.txt
django
psycopg2
dj_database_url
pillow
django-allauth